+++ Aps-Api/aps-monthplan-api/src/main/java/com/zlt/aps/monthplan/api/domain/entity/DpDemandPlan.java
@@ -184,7 +184,7 @@ public class DpDemandPlan extends BaseEntity {
     @Excel(name = "ui.data.column.dpDemandPlan.orderQty", cellType = Excel.ColumnType.NUMERIC)
     @ApiModelProperty(value = "订单量", name = "orderQty")
     @TableField(value = "ORDER_QTY")
-    private BigDecimal orderQty;
+    private Integer orderQty;
 
     /**
      * 库存
@@ -192,7 +192,7 @@ public class DpDemandPlan extends BaseEntity {
     @Excel(name = "ui.data.column.demandPlan.stockQty", cellType = Excel.ColumnType.NUMERIC)
     @ApiModelProperty(value = "库存", name = "stockQty")
     @TableField(value = "STOCK_QTY")
-    private BigDecimal stockQty;
+    private Integer stockQty;
 
     /**
      * 月结库存余量
@@ -200,7 +200,7 @@ public class DpDemandPlan extends BaseEntity {
     @Excel(name = "ui.data.column.demandPlan.remainingQty", cellType = Excel.ColumnType.NUMERIC)
     @ApiModelProperty(value = "月结库存余量", name = "remainingQty")
     @TableField(value = "REMAINING_QTY")
-    private BigDecimal remainingQty;
+    private Integer remainingQty;
 
     /**
      * 月均销量
@@ -208,7 +208,7 @@ public class DpDemandPlan extends BaseEntity {
     @Excel(name = "ui.data.column.dpDemandPlan.averageSaleQty", cellType = Excel.ColumnType.NUMERIC)
     @ApiModelProperty(value = "月均销量", name = "averageSaleQty")
     @TableField(value = "AVERAGE_SALE_QTY")
-    private BigDecimal averageSaleQty;
+    private Integer averageSaleQty;
 
     /**
      * 月底余量
@@ -216,7 +216,7 @@ public class DpDemandPlan extends BaseEntity {
     @Excel(name = "ui.data.column.demandPlan.plannedSurplus", cellType = Excel.ColumnType.NUMERIC)
     @ApiModelProperty(value = "月底余量", name = "plannedSurplus")
     @TableField(value = "PLANNED_SURPLUS")
-    private BigDecimal plannedSurplus;
+    private Integer plannedSurplus;
 
     /**
      * 排产净需求
@@ -224,7 +224,7 @@ public class DpDemandPlan extends BaseEntity {
     @Excel(name = "ui.data.column.demandPlan.netQty", cellType = Excel.ColumnType.NUMERIC)
     @ApiModelProperty(value = "排产净需求", name = "netQty")
     @TableField(value = "NET_QTY")
-    private BigDecimal netQty;
+    private Integer netQty;
 
     /**
      * 是否排产
@@ -240,7 +240,7 @@ public class DpDemandPlan extends BaseEntity {
     @Excel(name = "ui.data.column.demandPlan.postponeNetQty", cellType = Excel.ColumnType.NUMERIC)
     @ApiModelProperty(value = "净需求(含暂缓)", name = "postponeNetQty")
     @TableField(value = "POSTPONE_NET_QTY")
-    private BigDecimal postponeNetQty;
+    private Integer postponeNetQty;
 
     /**
      * 净需求(不含暂缓)
@@ -248,7 +248,7 @@ public class DpDemandPlan extends BaseEntity {
     @Excel(name = "ui.data.column.dpDemandPlan.unPostponeNetQty", cellType = Excel.ColumnType.NUMERIC)
     @ApiModelProperty(value = "净需求(不含暂缓)", name = "unPostponeNetQty")
     @TableField(value = "UN_POSTPONE_NET_QTY")
-    private BigDecimal unPostponeNetQty;
+    private Integer unPostponeNetQty;
 
     /**
      * 高优先级
@@ -256,7 +256,7 @@ public class DpDemandPlan extends BaseEntity {
     @Excel(name = "ui.data.column.demandPlan.heightQty", cellType = Excel.ColumnType.NUMERIC)
     @ApiModelProperty(value = "高优先级", name = "heightQty")
     @TableField(value = "HEIGHT_QTY")
-    private BigDecimal heightQty;
+    private Integer heightQty;
 
     /**
      * 中优先级
@@ -264,7 +264,7 @@ public class DpDemandPlan extends BaseEntity {
     @Excel(name = "ui.data.column.demandPlandemandPlan.midQty", cellType = Excel.ColumnType.NUMERIC)
     @ApiModelProperty(value = "中优先级", name = "midQty")
     @TableField(value = "MID_QTY")
-    private BigDecimal midQty;
+    private Integer midQty;
 
     /**
      * 暂缓订单
@@ -272,7 +272,7 @@ public class DpDemandPlan extends BaseEntity {
     @Excel(name = "ui.data.column.demandPlan.postponeQty", cellType = Excel.ColumnType.NUMERIC)
     @ApiModelProperty(value = "暂缓订单", name = "postponeQty")
     @TableField(value = "POSTPONE_QTY")
-    private BigDecimal postponeQty;
+    private Integer postponeQty;
 
     /**
      * 周期排产储备
@@ -280,7 +280,7 @@ public class DpDemandPlan extends BaseEntity {
     @Excel(name = "ui.data.column.demandPlan.cycleReserveQty", cellType = Excel.ColumnType.NUMERIC)
     @ApiModelProperty(value = "周期排产储备", name = "cycleReserveQty")
     @TableField(value = "CYCLE_RESERVE_QTY")
-    private BigDecimal cycleReserveQty;
+    private Integer cycleReserveQty;
 
     /**
      * 常规储备
@@ -288,7 +288,7 @@ public class DpDemandPlan extends BaseEntity {
     @Excel(name = "ui.data.column.demandPlan.conventionReserveQty", cellType = Excel.ColumnType.NUMERIC)
     @ApiModelProperty(value = "常规储备", name = "conventionReserveQty")
     @TableField(value = "CONVENTION_RESERVE_QTY")
-    private BigDecimal conventionReserveQty;
+    private Integer conventionReserveQty;
 
     /**
      * 是否满足最小投产量
@@ -304,7 +304,7 @@ public class DpDemandPlan extends BaseEntity {
     @Excel(name = "ui.data.column.demandPlan.minProductionQty", cellType = Excel.ColumnType.NUMERIC)
     @ApiModelProperty(value = "最小投产量值", name = "minProductionQty")
     @TableField(value = "MIN_PRODUCTION_QTY")
-    private Long minProductionQty;
+    private Integer minProductionQty;
 
     /**
      * 计划类型
